Наводим порядок

+++ app/src/main/java/com/example/notesapp/data/apiservice/ApiService.kt
@@ -4,7 +4,7 @@ import com.example.notesapp.data.database.entitys.Notes
 import kotlinx.coroutines.flow.StateFlow
 
 interface ApiService  {
-    suspend fun getAllNote(delayTime: Long): NoteResponse
+    suspend fun getAllNote(delayTime: Int): NoteResponse
     suspend fun addNote(note: Notes, delayTime: Int): Long
     suspend fun deleteNote(note: Notes, delayTime: Int)
     fun getChangeBaseTime(): Long
